requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
plotly
streamlit
python-dotenv
sqlalchemy
joblib
google-generativeai
